[bot] refs #7 - Move vote callback helpers to vote.py
