+++ bot/handlers/utils.py
@@ -25,28 +25,3 @@ def clean_config_data(data):
         del data['options']
     except KeyError:
         pass
-
-#Vote Callback helpers
-AADD = 1
-AREM = 2
-ACAN = 3
-ACOM = 4
-
-VOTE_SEL_REGEX = r'VOTESEL\*(.+)\*$'
-VOTE_SEL_PATTERN = 'VOTESEL*%s*'
-VOTE_REGEX = r'VOTE\*(.+)\*$'
-VOTE_PATTERN = 'VOTE*%s*'
-
-def vote_parse_cdata(cdata):
-    idx, typex, option = re.findall('^(.+):([1|2|3|4]):(.*)$', cdata)[0]
-    return int(idx), int(typex), option
-
-def vote_build_cdata(chat_id, option, typex):
-    return VOTE_PATTERN%(f'{chat_id}:{typex}:{option}')
-
-def vote_parse_selected(data: str):
-    parts = data.split(':')
-    if len(parts) > 1:
-        result = re.findall('([0-9]+) - (.*)', parts[-1])
-        return [ op for _, op in result ]
-    return []
